+++ scriptabit/__init__.py
@@ -11,6 +11,7 @@ from .configuration import (
     copy_default_config_to_user_directory,
 )
 from .habitica_service import HabiticaService
+from .iplugin import IPlugin
 from .scriptabit import start_cli
 from .utility_functions import UtilityFunctions