+++ scriptabit/errors.py
@@ -40,3 +40,13 @@ class ArgumentOutOfRangeError(Exception):
 
     def __str__(self):
         return repr(self.value)
+
+
+class PluginError(Exception):
+    """Plugin error"""
+
+    def __init__(self, value):
+        self.value = value
+
+    def __str__(self):
+        return repr(self.value)