+++ scriptabit/scriptabit.py
@@ -12,13 +12,18 @@ from builtins import *
 
 import logging
 import logging.config
+import os
+from time import sleep
+
+from pkg_resources import Requirement, resource_filename
+from yapsy.PluginManager import PluginManager
 
 from .authentication import load_authentication_credentials
 from .configuration import (
     get_configuration,
     get_config_file,
     copy_default_config_to_user_directory)
-from .errors import ServerUnreachableError
+from .errors import ServerUnreachableError, PluginError
 from .habitica_service import HabiticaService
 from .metadata import __version__
 from .utility_functions import UtilityFunctions
@@ -37,37 +42,108 @@ def __init_logging(logging_config_file):
 
     # Load the config
     logging.config.fileConfig(get_config_file(logging_config_file))
-    logging.getLogger(__name__).debug('Logging online')
 
-def __get_configuration():
+def __get_configuration(plugin_manager):
     """ Builds and parses the hierarchical configuration from environment
-variables, configuration files, command-line arguments, and argument defaults.
+    variables, configuration files, command-line arguments,
+    and argument defaults.
+
+    Args:
+        plugin_manager (yapsy.PluginManager): The plugin manager.
 
     Returns: The argparse compatible configuration object.
     """
 
     extra_args = [UtilityFunctions.get_arg_parser()]
-    # TODO: Give all plugins a chance to extend the command line args.
+
+    # Plugins can define additional arguments
+    for plugin_info in plugin_manager.getAllPlugins():
+        print(plugin_info.name)
+        plugin_arg_parser = plugin_info.plugin_object.get_arg_parser()
+        extra_args.append(plugin_arg_parser)
+
     config, _ = get_configuration(parents=extra_args)
 
     return config
 
+def __get_plugin_manager():
+    """ Discovers and instantiates all plugins, returning a management object.
+
+    Returns: yapsy.PluginManager: The plugin manager with the loaded plugins.
+    """
+
+    def init_user_plugin_directory():
+        """ Locates (and creates if necessary) the user plugin directory. """
+
+        default = os.path.expanduser('~/scriptabit_plugins')
+        user = os.path.expanduser(os.getenv('SCRIPTABIT_USER_PLUGIN_DIR', ''))
+        plugin_dir = user if user else default
+
+        if not os.path.exists(plugin_dir):
+            # Can't use logging, as logging is initialised after plugins
+            print('Creating user plugin directory {0}'.format(plugin_dir))
+            os.mkdir(plugin_dir)
+
+        return plugin_dir
+
+    # Build the manager
+    plugin_manager = PluginManager()
+
+    # the location of the plugins that ship with scriptabit
+    package_plugin_path = resource_filename(
+        Requirement.parse("scriptabit"),
+        os.path.join('scriptabit', 'plugins'))
+
+    # user plugin location
+    user_plugin_path = init_user_plugin_directory()
+
+    # Set plugin locations
+    plugin_manager.setPluginPlaces([package_plugin_path, user_plugin_path])
+
+    # Load all plugins
+    plugin_manager.collectPlugins()
+
+    return plugin_manager
+
+def __list_plugins(plugin_manager):
+    """ Lists the available plugins.
+
+    Args:
+        plugin_manager (yapsy.PluginManager): the plugin manager containing
+        the plugins.
+        """
+
+    def print_plugin_metadata(plugin_info):
+        """Utility class to pretty-print plugin information."""
+
+        print('* {0}:'.format(plugin_info.name))
+        print('{0}'.format(plugin_info.description))
+
+    print()
+    print('---- Plugins ----')
+    print('To execute a plugin, use the plugin name with the -p argument.')
+    print()
+    for plugin_info in plugin_manager.getAllPlugins():
+        print_plugin_metadata(plugin_info)
+        print()
+    print('-----------------')
+    print()
+
 def start_cli():
     """ Command-line entry point for scriptabit """
 
-    config = __get_configuration()
-
-    # TODO: build a list of available scenarios
+    plugin_manager = __get_plugin_manager()
+    config = __get_configuration(plugin_manager)
     __init_logging(config.logging_config)
-
     logging.getLogger(__name__).info('scriptabit version %s', __version__)
 
     # Disabling the broad exception warning as catching
     # everything is *exactly* the intent here.
     # pylint: disable=broad-except
     try:
-        if config.list_scenarios:
-            logging.getLogger(__name__).debug('Listing available scenarios')
+        if config.list_plugins:
+            logging.getLogger(__name__).debug('Listing available plugins')
+            __list_plugins(plugin_manager)
         else:
             # --------------------------------------------------
             # Running against Habitica.
@@ -96,12 +172,33 @@ def start_cli():
             utility = UtilityFunctions(config, habitica_service)
             utility.run()
 
-            if config.scenario:
-                # Time to run the selected scenario
-                logging.getLogger(__name__).debug(
-                    "Running '%s' scenario", config.scenario)
+            if config.plugin:
+                # Time to run the selected plugin
+                # First, find it
+                logging.getLogger(__name__).info("** %s running", config.plugin)
 
-                # TODO: scenario factory and execution
+                plugin_info = plugin_manager.getPluginByName(config.plugin)
+
+                if not plugin_info:
+                    raise PluginError('plugin %s not found' % config.plugin)
+
+                # Second, activate it
+                plugin_manager.activatePluginByName(config.plugin)
+                plugin = plugin_info.plugin_object
+                plugin.initialise(config, habitica_service)
+
+                # Finally, run it
+                updating = True
+                count = 0
+                while updating:
+                    logging.getLogger(__name__).debug("%s update %d",
+                                                      config.plugin,
+                                                      count)
+                    updating = plugin.update()
+                    count += 1
+                    sleep(plugin.update_interval_seconds())
+                logging.getLogger(__name__).info("** %s done", config.plugin)
+
     except Exception as exception:
         logging.getLogger(__name__).error(exception, exc_info=True)
         # pylint: enable=broad-except